feat: 이미지 캐시 처리 및 이미지 생성 분리
