+++ BoxOffice/Infra/Util/CacheManager.swift
@@ -1,0 +1,6 @@
+import UIKit
+
+class CacheManager {
+    static let shared = NSCache<NSString, UIImage>()
+    private init() {}
+}
